fix: support EmbeddedGUI TC32 integration

+++ example/HelloMultiDisplayHetero/uicode_disp0.c
@@ -77,7 +77,7 @@ void uicode_disp0_init(egui_core_t *core)
     egui_view_set_position(EGUI_VIEW_OF(&s_main_viewpage), 0, 0);
     egui_view_set_size(EGUI_VIEW_OF(&s_main_viewpage), EGUI_CONFIG_SCREEN_WIDTH, EGUI_CONFIG_SCREEN_HEIGHT);
 
-    for (i = 0; i < EGUI_ARRAY_SIZE(s_page_labels); i++)
+    for (i = 0; i < (int)EGUI_ARRAY_SIZE(s_page_labels); i++)
     {
         egui_view_label_init(EGUI_VIEW_OF(&s_page_labels[i]), core);
         egui_view_set_size(EGUI_VIEW_OF(&s_page_labels[i]), EGUI_CONFIG_SCREEN_WIDTH, EGUI_CONFIG_SCREEN_HEIGHT);
